+++ client/src/net/mod.rs
@@ -2,7 +2,6 @@ pub mod api_trait;
 pub mod error;
 pub mod http_client;
 pub mod protocol;
-pub mod tls;
 
 pub use api_trait::ApiClient;
 pub use error::ApiClientError;
