+++ server/Cargo.toml
@@ -18,6 +18,7 @@ clap = { workspace = true, features = ["string"] }
 libsignal-protocol = { workspace = true }
 derive_more = { workspace = true, features = ["display", "error", "from"] }
 sam-common = { workspace = true }
+sam-net = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 futures-util = { workspace = true, features = ["sink"] }
 async-trait = { workspace = true }
