+++ server/src/config.rs
@@ -1,8 +1,9 @@
 use std::net::SocketAddr;
 
+use sam_net::{error::ServerTlsError, tls::create_tls_server_config};
 use serde::{Deserialize, Serialize};
 
-use crate::{create_tls_config, error::TlsError, ServerState, StateType};
+use crate::{ServerState, StateType};
 
 pub struct ServerConfig<T: StateType> {
     pub state: ServerState<T>,
@@ -56,10 +57,10 @@ pub struct TlsConfig {
 impl TlsConfig {}
 
 impl TryInto<rustls::ServerConfig> for TlsConfig {
-    type Error = TlsError;
+    type Error = ServerTlsError;
 
     fn try_into(self) -> Result<rustls::ServerConfig, Self::Error> {
-        create_tls_config(
+        create_tls_server_config(
             &self.cert_path,
             &self.key_path,
             self.ca_cert_path.as_deref(),
