+++ server/src/error.rs
@@ -8,8 +8,9 @@ use axum::response::IntoResponse;
 use derive_more::derive::{Display, Error};
 use derive_more::From;
 use log::error;
-use rustls::server::VerifierBuilderError;
+
 use sam_common::LibError;
+use sam_net::error::ServerTlsError;
 
 pub type Result<T> = std::result::Result<T, ServerError>;
 
@@ -27,16 +28,8 @@ pub enum ServerError {
 }
 
 #[derive(Debug, Display, Error, From)]
-pub enum TlsError {
-    LoadError(std::io::Error),
-    VerifierError(VerifierBuilderError),
-    RustlsError(rustls::Error),
-    PrivateKeyWasNone,
-}
-
-#[derive(Debug, Display, Error, From)]
 pub enum CliError {
-    TLSError(TlsError),
+    TLSError(ServerTlsError),
     AddressParseError,
     SerdeError(serde_json::Error),
     IoError(std::io::Error),
