+++ server/src/lib.rs
@@ -7,7 +7,6 @@ pub mod protocol;
 pub mod routes;
 pub mod server;
 pub mod state;
-pub mod tls;
 
 #[cfg(test)]
 mod test_utils;
@@ -16,4 +15,3 @@ pub use config::ServerConfig;
 pub use error::ServerError;
 pub use server::start_server;
 pub use state::{state_type::StateType, ServerState};
-pub use tls::create_tls_config;
